+++ viscord-1.19.2-fabric-forge-template/common/src/main/java/network/vonix/viscord/discord/WebhookClient.java
@@ -42,7 +42,7 @@ public class WebhookClient {
     }
 
     public void sendMessage(String username, String avatarUrl, String content) {
-        if (webhookUrl == null || webhookUrl.isEmpty() || webhookUrl.isEmpty()) return;
+        if (webhookUrl == null || webhookUrl.isEmpty()) return;
 
         JsonObject json = new JsonObject();
         json.addProperty("username", username);